requirements.txt
docker
PyYAML
cltools
six
